[android][sdk] Fix dependecies in sdk package
