+++ android/app/src/main/java/app/organicmaps/MwmActivity.java
@@ -569,7 +569,7 @@ public class MwmActivity extends BaseMwmFragmentActivity
 
     mShareLauncher = SharingUtils.RegisterLauncher(this);
 
-    mDisplayManager = DisplayManager.from(this);
+    mDisplayManager = MwmApplication.from(this).getDisplayManager();
     if (mDisplayManager.isCarDisplayUsed())
     {
       mRemoveDisplayListener = false;
