+++ android/app/src/main/java/app/organicmaps/car/SurfaceRenderer.java
@@ -14,11 +14,11 @@ import androidx.car.app.SurfaceContainer;
 import androidx.lifecycle.DefaultLifecycleObserver;
 import androidx.lifecycle.Lifecycle;
 import androidx.lifecycle.LifecycleOwner;
+import app.organicmaps.MwmApplication;
 import app.organicmaps.R;
 import app.organicmaps.sdk.Framework;
 import app.organicmaps.sdk.Map;
 import app.organicmaps.sdk.MapRenderingListener;
-import app.organicmaps.sdk.display.DisplayManager;
 import app.organicmaps.sdk.settings.UnitLocale;
 import app.organicmaps.sdk.util.concurrency.UiThread;
 import app.organicmaps.sdk.util.log.Logger;
@@ -116,7 +116,7 @@ public class SurfaceRenderer implements DefaultLifecycleObserver, SurfaceCallbac
   {
     Logger.d(TAG);
     mMap.onResume();
-    if (DisplayManager.from(mCarContext).isCarDisplayUsed())
+    if (MwmApplication.from(mCarContext).getDisplayManager().isCarDisplayUsed())
       UiThread.runLater(() -> mMap.updateMyPositionRoutingOffset(0));
   }
