+++ android/app/src/main/java/app/organicmaps/editor/data/TimeFormatUtils.java
@@ -1,9 +1,11 @@
-package app.organicmaps.sdk.editor.data;
+package app.organicmaps.editor.data;
 
 import android.content.res.Resources;
 import androidx.annotation.IntRange;
 import androidx.annotation.NonNull;
 import app.organicmaps.R;
+import app.organicmaps.sdk.editor.data.Timespan;
+import app.organicmaps.sdk.editor.data.Timetable;
 import app.organicmaps.util.Utils;
 import java.text.DateFormatSymbols;
 import java.util.Locale;
