+++ android/app/src/main/java/app/organicmaps/sdk/display/DisplayManager.java
@@ -1,11 +1,9 @@
 package app.organicmaps.sdk.display;
 
-import android.content.Context;
 import android.os.Handler;
 import android.os.Looper;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import app.organicmaps.MwmApplication;
 import app.organicmaps.sdk.util.log.Logger;
 import java.util.Objects;
 
@@ -38,13 +36,6 @@ public class DisplayManager
   private DisplayHolder mDevice;
   @Nullable
   private DisplayHolder mCar;
-
-  @NonNull
-  public static DisplayManager from(@NonNull Context context)
-  {
-    final MwmApplication app = (MwmApplication) context.getApplicationContext();
-    return app.getDisplayManager();
-  }
 
   public boolean isCarConnected()
   {
