+++ android/app/src/main/java/app/organicmaps/sdk/downloader/ChunkTask.java
@@ -6,8 +6,8 @@ import androidx.annotation.Keep;
 import app.organicmaps.downloader.Android7RootCertificateWorkaround;
 import app.organicmaps.sdk.util.Constants;
 import app.organicmaps.sdk.util.StringUtils;
+import app.organicmaps.sdk.util.Utils;
 import app.organicmaps.sdk.util.log.Logger;
-import app.organicmaps.util.Utils;
 import java.io.BufferedInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
