+++ android/app/src/main/java/app/organicmaps/sdk/editor/OpeningHours.java
@@ -25,12 +25,6 @@ public final class OpeningHours
   public static native Timetable nativeGetComplementTimetable(Timetable[] timetableSet);
 
   @NonNull
-  public static native Timetable[] nativeAddTimetable(Timetable[] timetableSet);
-
-  @NonNull
-  public static native Timetable[] nativeRemoveTimetable(Timetable[] timetableSet, int timetableIndex);
-
-  @NonNull
   public static native Timetable nativeSetIsFullday(Timetable timetable, boolean isFullday);
 
   @NonNull
