+++ android/app/src/main/java/app/organicmaps/search/SearchFragment.java
@@ -176,7 +176,7 @@ public class SearchFragment extends BaseMwmFragment implements SearchListener, C
 
   private final LocationListener mLocationListener = new LocationListener() {
     @Override
-    public void onLocationUpdated(Location location)
+    public void onLocationUpdated(@NonNull Location location)
     {
       mLastPosition.set(location.getLatitude(), location.getLongitude());
